# docs/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/items/{id}/studio": {
      "get": {
        "operationId": "StudioController_studio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "商品工作区"
        ]
      }
    },
    "/api/items/{id}/review-for-use": {
      "post": {
        "operationId": "StudioController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "商品工作区"
        ]
      }
    },
...
    "/api/channels/{id}": {
      "post": {
        "operationId": "PublishingController_updateChannel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
    "/api/items/{id}/readiness": {
      "get": {
        "operationId": "PublishingController_readiness",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
    "/api/items/{id}/prepare": {
      "post": {
        "operationId": "PublishingController_prepare",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
    "/api/items/{id}/package-preview": {
      "get": {
        "operationId": "PublishingController_preview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
    "/api/items/{id}/packages": {
      "post": {
        "operationId": "PublishingController_pack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
    "/api/packages/{id}": {
      "get": {
        "operationId": "PublishingController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
    "/api/packages/{id}/download": {
      "get": {
        "operationId": "PublishingController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
...
    "/api/listings/{id}/observe": {
      "post": {
        "operationId": "PublishingController_observe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
...
    "/api/showroom/{id}/image/{asset}": {
      "get": {
        "operationId": "PublishingController_publicImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asset",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "使用与分发"
        ]
      }
    },
...
  },
  "info": {
    "title": "ToMeBoutique Workbench API",
    "description": "",
    "version": "1.0.0-rc.14",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tome_session"
      }
    },
...
  }
}